feat: 관람자 중도 참여시 턴 초기값 설정 및 턴 변경시 턴 카운트 업데이트 구현
